+++ sites/public/lib/helpers.ts
@@ -1,1 +1,5 @@
+import { ELIGIBILITY_ROUTE, ELIGIBILITY_SECTIONS } from "./constants"
+
 export const emailRegex = /^(([^<>()[\]\\.,;:\s@"]+(\.[^<>()[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/
+export const eligibilityRoute = (page: number) =>
+  `/${ELIGIBILITY_ROUTE}/${ELIGIBILITY_SECTIONS[page]}`