+++ sites/public/pages/eligibility/disability.tsx
@@ -13,18 +13,32 @@ import {
 } from "@bloom-housing/ui-components"
 import FormsLayout from "../../layouts/forms"
 import { useForm } from "react-hook-form"
-import React from "react"
+import React, { useContext } from "react"
 import { useRouter } from "next/router"
 import { ELIGIBILITY_ROUTE, ELIGIBILITY_SECTIONS } from "../../lib/constants"
+import { EligibilityContext } from "../../lib/EligibilityContext"
+import { eligibilityRoute } from "../../lib/helpers"
+import FormBackLink from "../../src/forms/applications/FormBackLink"
 
 const EligibilityDisability = () => {
   const router = useRouter()
+  const { eligibilityRequirements } = useContext(EligibilityContext)
+  const CURRENT_PAGE = 3
 
   /* Form Handler */
   // eslint-disable-next-line @typescript-eslint/unbound-method
-  const { handleSubmit, register, errors } = useForm()
+  const { handleSubmit, register, errors, getValues } = useForm({
+    defaultValues: {
+      disability: eligibilityRequirements?.disability,
+    },
+  })
+
   const onSubmit = () => {
-    void router.push(`/${ELIGIBILITY_ROUTE}/${ELIGIBILITY_SECTIONS[4]}`)
+    const data = getValues()
+    const { disability } = data
+    eligibilityRequirements.setDisability(disability)
+
+    void router.push(eligibilityRoute(CURRENT_PAGE + 1))
   }
 
   const disabilityValues = [
@@ -55,6 +69,12 @@ const EligibilityDisability = () => {
         />
       </FormCard>
       <FormCard>
+        <FormBackLink
+          url={eligibilityRoute(CURRENT_PAGE - 1)}
+          onClick={() => {
+            // Not extra actions needed.
+          }}
+        />
         <Form onSubmit={handleSubmit(onSubmit)}>
           <div className="form-card__lead pb-0 pt-8">
             <h2 className="form-card__title is-borderless">{t("eligibility.disability.prompt")}</h2>
@@ -66,7 +86,7 @@ const EligibilityDisability = () => {
               <FieldGroup
                 type="radio"
                 name="disability"
-                error={errors.disability}
+                error={errors.disability != null}
                 errorMessage={t("errors.selectOption")}
                 register={register}
                 validation={{ required: true }}