+++ sites/public/pages/eligibility/income.tsx
@@ -3,7 +3,7 @@ Income
 Prompts the user for their annual income.
 */
 import FormsLayout from "../../layouts/forms"
-import React from "react"
+import React, { useContext } from "react"
 import { FormCard } from "@bloom-housing/ui-components/src/blocks/FormCard"
 import { t } from "@bloom-housing/ui-components/src/helpers/translator"
 import { ProgressNav } from "@bloom-housing/ui-components/src/navigation/ProgressNav"
@@ -13,17 +13,30 @@ import { Button } from "@bloom-housing/ui-components/src/actions/Button"
 import { AppearanceStyleType, Select } from "@bloom-housing/ui-components"
 import { useRouter } from "next/router"
 import { useForm } from "react-hook-form"
+import { EligibilityContext } from "../../lib/EligibilityContext"
+import { eligibilityRoute } from "../../lib/helpers"
+import FormBackLink from "../../src/forms/applications/FormBackLink"
 
 const EligibilityIncome = () => {
   const router = useRouter()
+  const { eligibilityRequirements } = useContext(EligibilityContext)
 
   const incomeRanges = ["below10k", "10kTo20k", "30kTo40k", "40kTo50k", "over50k"]
+  const CURRENT_PAGE = 4
 
   /* Form Handler */
   // eslint-disable-next-line @typescript-eslint/unbound-method
-  const { handleSubmit, register } = useForm()
+  const { handleSubmit, register, getValues } = useForm({
+    defaultValues: {
+      income: eligibilityRequirements?.income ?? incomeRanges[0],
+    },
+  })
   const onSubmit = () => {
-    void router.push(`/${ELIGIBILITY_ROUTE}/${ELIGIBILITY_SECTIONS[5]}`)
+    const data = getValues()
+    const { income } = data
+    eligibilityRequirements.setIncome(income)
+
+    void router.push(eligibilityRoute(CURRENT_PAGE + 1))
   }
 
   return (
@@ -36,6 +49,12 @@ const EligibilityIncome = () => {
         />
       </FormCard>
       <FormCard>
+        <FormBackLink
+          url={eligibilityRoute(CURRENT_PAGE - 1)}
+          onClick={() => {
+            // Not extra actions needed.
+          }}
+        />
         <div className="form-card__lead pb-0 pt-8">
           <h2 className="form-card__title is-borderless">{t("eligibility.income.prompt")}</h2>
         </div>
@@ -45,12 +64,11 @@ const EligibilityIncome = () => {
               {t("eligibility.income.description")}
             </p>
             <Select
-              id="eligibility.income"
-              name="eligibility.income"
+              id="income"
+              name="income"
               label={t("eligibility.income.label")}
               describedBy="income-description"
               validation={{ required: true }}
-              defaultValue={t("eligibility.income.ranges.below10k")}
               register={register}
               controlClassName="control"
               options={incomeRanges}