+++ sites/public/pages/eligibility/section8.tsx
@@ -3,7 +3,7 @@ Section 8
 Asks whether the user has a section 8 voucher.
 */
 import FormsLayout from "../../layouts/forms"
-import React from "react"
+import React, { useContext } from "react"
 import { FormCard } from "@bloom-housing/ui-components/src/blocks/FormCard"
 import { t } from "@bloom-housing/ui-components/src/helpers/translator"
 import { ProgressNav } from "@bloom-housing/ui-components/src/navigation/ProgressNav"
@@ -12,13 +12,25 @@ import { Form } from "@bloom-housing/ui-components/src/forms/Form"
 import { Button } from "@bloom-housing/ui-components/src/actions/Button"
 import { AppearanceStyleType, FieldGroup } from "@bloom-housing/ui-components"
 import { useForm } from "react-hook-form"
+import { EligibilityContext } from "../../lib/EligibilityContext"
+import FormBackLink from "../../src/forms/applications/FormBackLink"
+import { eligibilityRoute } from "../../lib/helpers"
 
 const EligibilitySection8 = () => {
+  const { eligibilityRequirements } = useContext(EligibilityContext)
+  const CURRENT_PAGE = 5
+
   /* Form Handler */
   // eslint-disable-next-line @typescript-eslint/unbound-method
-  const { handleSubmit, register, errors } = useForm()
+  const { handleSubmit, register, errors, getValues } = useForm({
+    defaultValues: {
+      section8: eligibilityRequirements?.section8,
+    },
+  })
   const onSubmit = () => {
-    // Not yet implemented.
+    const data = getValues()
+    const { section8 } = data
+    eligibilityRequirements.setSection8(section8)
   }
 
   const section8Values = [
@@ -44,6 +56,12 @@ const EligibilitySection8 = () => {
         />
       </FormCard>
       <FormCard>
+        <FormBackLink
+          url={eligibilityRoute(CURRENT_PAGE - 1)}
+          onClick={() => {
+            // Not extra actions needed.
+          }}
+        />
         <Form onSubmit={handleSubmit(onSubmit)}>
           <div className="form-card__lead pb-0 pt-8">
             <h2 className="form-card__title is-borderless">{t("eligibility.section8.prompt")}</h2>
@@ -54,7 +72,7 @@ const EligibilitySection8 = () => {
               <FieldGroup
                 type="radio"
                 name="section8"
-                error={errors.section8}
+                error={errors.section8 != null}
                 errorMessage={t("errors.selectOption")}
                 register={register}
                 validation={{ required: true }}