+++ sites/public/pages/eligibility/welcome.tsx
@@ -15,14 +15,16 @@ import { useForm } from "react-hook-form"
 import React from "react"
 import { useRouter } from "next/router"
 import { ELIGIBILITY_ROUTE, ELIGIBILITY_SECTIONS } from "../../lib/constants"
+import { eligibilityRoute } from "../../lib/helpers"
 
 const EligibilityWelcome = () => {
   const router = useRouter()
+  const CURRENT_PAGE = 0
 
   /* Form Handler */
   const { handleSubmit } = useForm()
   const onSubmit = () => {
-    void router.push(`/${ELIGIBILITY_ROUTE}/${ELIGIBILITY_SECTIONS[1]}`)
+    void router.push(eligibilityRoute(CURRENT_PAGE + 1))
   }
 
   return (